Change how per-environment hosts are read
- use Rakefile to pass environment, not host
- improve Rakefile
- don't index non production environment sites
