+++ config.rb
@@ -3,17 +3,33 @@ require 'bundler'
 
 Bundler.require(:default, config[:environment], config[:mode])
 
-config[:host] = ENV['HOST']
+# Add project top-level directory to load path
+$LOAD_PATH << File.expand_path('../', __FILE__)
+require 'lib/helpers'
 
 ###
-# Configure Middleman
+# Global Configuration
 ###
+
+# Add all Helper module submodules as helpers
+# both throughout app and in this config file
+Helpers.constants(false)
+       .map    { |c| Helpers.const_get(c) }
+       .select { |c| c.is_a? Module }
+       .each do |c|
+  helpers c
+  include c
+end
+
+# Do not render with layout
+page '/*.txt', layout: false
+page '/*.xml', layout: false
 
 # Asset pipeline
 activate :sprockets
 sprockets.append_path File.join(root, 'source')
 
-# CSS
+# CSS vendor prefixes
 activate :autoprefixer do |config|
   config.browsers = ['last 2 versions', 'Explorer >= 9']
   config.remove   = false
@@ -21,39 +37,22 @@ activate :autoprefixer do |config|
   config.inline   = true
 end
 
-# Do not render with layout
-page '/*.txt', layout: false
-page '/*.xml', layout: false
+###
+# Configuration by environment
+###
 
-# Helpers
-require 'helpers/google_helpers'
-require 'helpers/gravatar_helpers'
-require 'helpers/resume_helpers'
-
-helpers GoogleHelpers
-helpers GravatarHelpers
-helpers ResumeHelpers
-
-# Deploy configuration
-
-unless config[:host].nil?
-  # TODO: shouldn't have to run this other than during deploys
-  # NB: Reads credentials from .s3_sync; assumes same credentials used for all environments
-  activate :s3_sync do |s3_sync|
-    s3_sync.bucket                     = config[:host].sub(%r{^https?\:\/\/}, '')
-    s3_sync.region                     = 'us-west-1'
-    s3_sync.prefer_gzip                = true
-    s3_sync.index_document             = 'index.html'
-    s3_sync.error_document             = 'error.html'
-  end
-end
-
-# General configuration
+config[:host] = app.data[:hosts][config[:environment]]
+config_s3 host: config[:host]
 
 configure :development do
   # reload the browser automatically whenever files change
   activate :livereload
 end
+
+###
+# Configuration by mode
+# mdzhang: You would think there would be a :deploy mode
+###
 
 configure :build do
   activate :gzip
