--- helpers/resume_helpers.rb
@@ -1,5 +1,0 @@
-module ResumeHelpers
-  def resume_url
-    "#{config[:host]}/files/#{data.me.resume_file_name}"
-  end
-end
