+++ lib/helpers.rb
@@ -1,0 +1,7 @@
+module Helpers
+end
+
+require 'lib/helpers/deploy'
+require 'lib/helpers/google'
+require 'lib/helpers/gravatar'
+require 'lib/helpers/resume'
