+++ source/layouts/layout.html.haml
@@ -1,7 +1,6 @@
 !!!
 %html{ lang: 'en' }
+  = partial 'partials/head'
 
-= partial 'partials/head'
-
-%body
-  = yield
+  %body
+    = yield
