+++ source/robots.txt.haml
@@ -1,0 +1,3 @@
+- unless config[:environment].eql?(:production)
+  User-agent: *
+  Disallow: /
